Revert "EVIV screen revamp"
This reverts commit 9e3e0e25684ada92b17d8bc3b4c21a7c17d81293.

+++ src/main_eviv.h
@@ -54,7 +54,7 @@ struct EvIvDisplayScreen
  
     u8 stats_ev[NUM_STATS];
     u8 stats_iv[NUM_STATS];
-    u16 stats_bs[NUM_STATS];
+    u8 stats_bs[NUM_STATS];
     u16 totalStatsEV;
     u16 totalStatsIV;
     u16 totalStatsBS;
